refactor(gateway): extract DaoHeaderResolver (#106 phase 2)
Second half of the #106 split. Pulls the three DAO chain-state
lookups out of GatewayRepository into their own @Singleton service.
GatewayRepository drops from 2375 to 2292 lines.

Moved to DaoHeaderResolver:
- getBlockHashForCell (light-client cache → peer fetch retry loop)
- fetchHeaderWithRetry (direct peer-fetch with retry)
- getOrFetchHeader (Room cache → local JNI → peer fetch tier)

The helpers only needed Json + DaoSyncManager — no per-call state to
thread through. getOrFetchHeader takes NetworkType as a parameter so
the resolver itself stays network-agnostic (caller passes
currentNetwork at the GatewayRepository delegation shim).

GatewayRepository keeps two thin shims (getBlockHashForCell and
getOrFetchHeader without the network arg) so internal call sites
inside getDaoDeposits / withdrawFromDao / unlockDao remain unchanged.

Phase 1 + phase 2 combined: GatewayRepository.kt 2629 → 2292 lines.
DI surface stable; ViewModels and existing tests untouched.

611/611 tests pass.

Refs #106
