+++ android/app/src/main/java/com/rjnr/pocketnode/di/AppModule.kt
@@ -241,5 +241,6 @@ object AppModule {
         pendingBroadcastDao: PendingBroadcastDao,
         broadcastClient: BroadcastClient,
         syncCoordinator: com.rjnr.pocketnode.data.gateway.SyncCoordinator,
-    ): GatewayRepository = GatewayRepository(context, keyManager, walletPreferences, json, transactionBuilder, cacheManager, daoSyncManager, walletMigrationHelper, walletDao, appDatabase, headerCacheDao, syncProgressDao, pendingBroadcastDao, broadcastClient, syncCoordinator)
+        daoHeaderResolver: com.rjnr.pocketnode.data.gateway.DaoHeaderResolver,
+    ): GatewayRepository = GatewayRepository(context, keyManager, walletPreferences, json, transactionBuilder, cacheManager, daoSyncManager, walletMigrationHelper, walletDao, appDatabase, headerCacheDao, syncProgressDao, pendingBroadcastDao, broadcastClient, syncCoordinator, daoHeaderResolver)
 }
